+++ src/estimation/mod.rs
@@ -1,3 +1,4 @@
+pub mod agq;
 pub mod bayes;
 pub mod gauss_newton;
 pub(crate) mod hmc;